Change computation 'Run' button functionality into props two props.
This moves towards only showing a 'Run Computation' button for a consortium's
owners by uses two props: `allowComputationRun` (disabled) and
`showComputationRunButton` (display the button). The 'show' prop should only
be `true` for a consortium's owner, and the 'allow' prop should only be true
if there isn't an active computation. Functionality behind the UI is found in
coinstac-client-core's computation service.