Rename `Pipeline#tryNext` → `Pipeline#maybeIncrementStep`.
This renames the `Pipeline` class's `tryNext` method to the more
appropriate name `maybeIncrementStep` as it contains the only code that
increments the pipeline's step property and progresses its computation
property. The method's return value is changed from an object
representing whether the pipeline incremented to a more appropriate
boolean value.